Split hosted operation validation modules
