+++ crates/connect-hosted-provider/src/provider/mutation_journal.rs
@@ -51,4 +51,5 @@ pub(super) enum HostedMutationClaim {
 }
 
 include!("mutation_journal/provider_impl.rs");
+include!("mutation_journal/terminal_replay.rs");
 include!("mutation_journal/helpers.rs");
